lib: avoidance: move table borders definition into applications
Table size is relative to each cup, so it must be declared in applications.
