+++ applications/cup2021/include/app.hpp
@@ -1,5 +1,8 @@
 #pragma once
 
+// Project includes
+#include "obstacles.hpp"
+
 #include "sd21.h"
 
 /*
@@ -110,6 +113,10 @@ static const sd21_conf_t sd21_config_app[] = {
         },
     },
 };
+
+/// @brief Return a polygon obstacle delimitng the table borders.
+/// @return                      table borders
+const cogip::obstacles::Polygon *app_get_borders(void);
 
 void app_init(void);
 void app_init_tasks(void);
